[refactor] signup 폴더명을 sign-up으로 변경
- 폴더 네이밍 컨벤션 kebab-case 통일
